feat: Bluetooth implementation
This is the first draft for bluetooth implementation.
once starting the UI, add goto localhost:8080/setup to test this
function.
It is NOT near ready for merge, just initial debugging for backend
developers to help debug the stream between Edge and UI.

One test in navbar had to be changed as we now have six items in the
menu. some tests have been written for bluetooth component.

This component will be rewritten in the future when we know how to make
one component handling all setup. This is just for now.